requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# Vector Database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2

# Web Interface
streamlit==1.29.0

# Data Processing
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# Optional: For local LLM (if needed)
# transformers==4.36.2
# torch==2.1.2